refactor(levm): clean main execution loop (#1824)
**Motivation**

The main execution loop has a lot of logic inside it that makes it hard
to follow the execution flow. This PR aims to make the execution loop
cleaner.

**Description**

Move logic away from the execute loop into to auxiliary functions. 

Closes #1825
