+++ crates/vm/levm/src/lib.rs
@@ -4,6 +4,7 @@ pub mod constants;
 pub mod db;
 pub mod environment;
 pub mod errors;
+pub mod execution_handlers;
 pub mod gas_cost;
 pub mod memory;
 pub mod opcode_handlers;
