+++ crates/vm/levm/src/opcode_handlers/block.rs
@@ -1,7 +1,7 @@
 use crate::{
     call_frame::CallFrame,
     constants::LAST_AVAILABLE_BLOCK_LIMIT,
-    errors::{InternalError, OpcodeSuccess, VMError},
+    errors::{InternalError, OpcodeResult, VMError},
     gas_cost,
     utils::*,
     vm::VM,
@@ -19,7 +19,7 @@ impl VM {
     pub fn op_blockhash(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         self.increase_consumed_gas(current_call_frame, gas_cost::BLOCKHASH)?;
 
         let block_number = current_call_frame.stack.pop()?;
@@ -33,7 +33,7 @@ impl VM {
             || block_number >= self.env.block_number
         {
             current_call_frame.stack.push(U256::zero())?;
-            return Ok(OpcodeSuccess::Continue);
+            return Ok(OpcodeResult::Continue);
         }
 
         let block_number: u64 = block_number
@@ -48,52 +48,52 @@ impl VM {
             current_call_frame.stack.push(U256::zero())?;
         }
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // COINBASE operation
     pub fn op_coinbase(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         self.increase_consumed_gas(current_call_frame, gas_cost::COINBASE)?;
 
         current_call_frame
             .stack
             .push(address_to_word(self.env.coinbase))?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // TIMESTAMP operation
     pub fn op_timestamp(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         self.increase_consumed_gas(current_call_frame, gas_cost::TIMESTAMP)?;
 
         current_call_frame.stack.push(self.env.timestamp)?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // NUMBER operation
     pub fn op_number(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         self.increase_consumed_gas(current_call_frame, gas_cost::NUMBER)?;
 
         current_call_frame.stack.push(self.env.block_number)?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // PREVRANDAO operation
     pub fn op_prevrandao(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         self.increase_consumed_gas(current_call_frame, gas_cost::PREVRANDAO)?;
 
         let randao = self.env.prev_randao.unwrap_or_default(); // Assuming block_env has been integrated
@@ -101,40 +101,40 @@ impl VM {
             .stack
             .push(U256::from_big_endian(randao.0.as_slice()))?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // GASLIMIT operation
     pub fn op_gaslimit(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         self.increase_consumed_gas(current_call_frame, gas_cost::GASLIMIT)?;
 
         current_call_frame
             .stack
             .push(self.env.block_gas_limit.into())?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // CHAINID operation
     pub fn op_chainid(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         self.increase_consumed_gas(current_call_frame, gas_cost::CHAINID)?;
 
         current_call_frame.stack.push(self.env.chain_id)?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // SELFBALANCE operation
     pub fn op_selfbalance(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         self.increase_consumed_gas(current_call_frame, gas_cost::SELFBALANCE)?;
 
         let balance = get_account(&mut self.cache, &self.db, current_call_frame.to)
@@ -142,19 +142,19 @@ impl VM {
             .balance;
 
         current_call_frame.stack.push(balance)?;
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // BASEFEE operation
     pub fn op_basefee(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         self.increase_consumed_gas(current_call_frame, gas_cost::BASEFEE)?;
 
         current_call_frame.stack.push(self.env.base_fee_per_gas)?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // BLOBHASH operation
@@ -162,7 +162,7 @@ impl VM {
     pub fn op_blobhash(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         // [EIP-4844] - BLOBHASH is only available from CANCUN
         if self.env.fork < Fork::Cancun {
             return Err(VMError::InvalidOpcode);
@@ -175,7 +175,7 @@ impl VM {
         let blob_hashes = &self.env.tx_blob_hashes;
         if index >= blob_hashes.len().into() {
             current_call_frame.stack.push(U256::zero())?;
-            return Ok(OpcodeSuccess::Continue);
+            return Ok(OpcodeResult::Continue);
         }
 
         let index: usize = index
@@ -191,7 +191,7 @@ impl VM {
             .stack
             .push(U256::from_big_endian(blob_hash.as_bytes()))?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     fn get_blob_gasprice(&mut self) -> Result<U256, VMError> {
@@ -209,7 +209,7 @@ impl VM {
     pub fn op_blobbasefee(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         // [EIP-7516] - BLOBBASEFEE is only available from CANCUN
         if self.env.fork < Fork::Cancun {
             return Err(VMError::InvalidOpcode);
@@ -220,7 +220,7 @@ impl VM {
 
         current_call_frame.stack.push(blob_base_fee)?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 }
