+++ crates/vm/levm/src/opcode_handlers/environment.rs
@@ -1,7 +1,7 @@
 use crate::{
     call_frame::CallFrame,
     constants::SET_CODE_DELEGATION_BYTES,
-    errors::{InternalError, OpcodeSuccess, VMError},
+    errors::{InternalError, OpcodeResult, VMError},
     gas_cost::{self},
     memory::{self, calculate_memory_size},
     utils::{has_delegation, word_to_address},
@@ -18,7 +18,7 @@ impl VM {
     pub fn op_address(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         self.increase_consumed_gas(current_call_frame, gas_cost::ADDRESS)?;
 
         let addr = current_call_frame.to; // The recipient of the current call.
@@ -27,14 +27,14 @@ impl VM {
             .stack
             .push(U256::from_big_endian(addr.as_bytes()))?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // BALANCE operation
     pub fn op_balance(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         let address = word_to_address(current_call_frame.stack.pop()?);
 
         let (account_info, address_was_cold) = self.access_account(address);
@@ -43,14 +43,14 @@ impl VM {
 
         current_call_frame.stack.push(account_info.balance)?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // ORIGIN operation
     pub fn op_origin(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         self.increase_consumed_gas(current_call_frame, gas_cost::ORIGIN)?;
 
         let origin = self.env.origin;
@@ -58,14 +58,14 @@ impl VM {
             .stack
             .push(U256::from_big_endian(origin.as_bytes()))?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // CALLER operation
     pub fn op_caller(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         self.increase_consumed_gas(current_call_frame, gas_cost::CALLER)?;
 
         let caller = current_call_frame.msg_sender;
@@ -73,28 +73,28 @@ impl VM {
             .stack
             .push(U256::from_big_endian(caller.as_bytes()))?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // CALLVALUE operation
     pub fn op_callvalue(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         self.increase_consumed_gas(current_call_frame, gas_cost::CALLVALUE)?;
 
         let callvalue = current_call_frame.msg_value;
 
         current_call_frame.stack.push(callvalue)?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // CALLDATALOAD operation
     pub fn op_calldataload(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         self.increase_consumed_gas(current_call_frame, gas_cost::CALLDATALOAD)?;
 
         let calldata_size: U256 = current_call_frame.calldata.len().into();
@@ -105,7 +105,7 @@ impl VM {
         // have no data to return.
         if offset > calldata_size {
             current_call_frame.stack.push(U256::zero())?;
-            return Ok(OpcodeSuccess::Continue);
+            return Ok(OpcodeResult::Continue);
         };
         let offset: usize = offset
             .try_into()
@@ -128,28 +128,28 @@ impl VM {
 
         current_call_frame.stack.push(result)?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // CALLDATASIZE operation
     pub fn op_calldatasize(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         self.increase_consumed_gas(current_call_frame, gas_cost::CALLDATASIZE)?;
 
         current_call_frame
             .stack
             .push(U256::from(current_call_frame.calldata.len()))?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // CALLDATACOPY operation
     pub fn op_calldatacopy(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         let dest_offset = current_call_frame.stack.pop()?;
         let calldata_offset = current_call_frame.stack.pop()?;
         let size: usize = current_call_frame
@@ -166,13 +166,13 @@ impl VM {
         )?;
 
         if size == 0 {
-            return Ok(OpcodeSuccess::Continue);
+            return Ok(OpcodeResult::Continue);
         }
 
         let mut data = vec![0u8; size];
         if calldata_offset > current_call_frame.calldata.len().into() {
             memory::try_store_data(&mut current_call_frame.memory, dest_offset, &data)?;
-            return Ok(OpcodeSuccess::Continue);
+            return Ok(OpcodeResult::Continue);
         }
 
         let calldata_offset: usize = calldata_offset
@@ -193,28 +193,28 @@ impl VM {
 
         memory::try_store_data(&mut current_call_frame.memory, dest_offset, &data)?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // CODESIZE operation
     pub fn op_codesize(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         self.increase_consumed_gas(current_call_frame, gas_cost::CODESIZE)?;
 
         current_call_frame
             .stack
             .push(U256::from(current_call_frame.bytecode.len()))?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // CODECOPY operation
     pub fn op_codecopy(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         let destination_offset = current_call_frame.stack.pop()?;
 
         let code_offset = current_call_frame.stack.pop()?;
@@ -233,7 +233,7 @@ impl VM {
         )?;
 
         if size == 0 {
-            return Ok(OpcodeSuccess::Continue);
+            return Ok(OpcodeResult::Continue);
         }
 
         let mut data = vec![0u8; size];
@@ -257,26 +257,26 @@ impl VM {
 
         memory::try_store_data(&mut current_call_frame.memory, destination_offset, &data)?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // GASPRICE operation
     pub fn op_gasprice(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         self.increase_consumed_gas(current_call_frame, gas_cost::GASPRICE)?;
 
         current_call_frame.stack.push(self.env.gas_price)?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // EXTCODESIZE operation
     pub fn op_extcodesize(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         let address = word_to_address(current_call_frame.stack.pop()?);
 
         let (account_info, address_was_cold) = self.access_account(address);
@@ -292,14 +292,14 @@ impl VM {
             account_info.bytecode.len().into()
         })?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // EXTCODECOPY operation
     pub fn op_extcodecopy(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         let address = word_to_address(current_call_frame.stack.pop()?);
         let dest_offset = current_call_frame.stack.pop()?;
         let offset = current_call_frame.stack.pop()?;
@@ -327,7 +327,7 @@ impl VM {
         )?;
 
         if size == 0 {
-            return Ok(OpcodeSuccess::Continue);
+            return Ok(OpcodeResult::Continue);
         }
 
         let bytecode = if is_delegation {
@@ -350,28 +350,28 @@ impl VM {
 
         memory::try_store_data(&mut current_call_frame.memory, dest_offset, &data)?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // RETURNDATASIZE operation
     pub fn op_returndatasize(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         self.increase_consumed_gas(current_call_frame, gas_cost::RETURNDATASIZE)?;
 
         current_call_frame
             .stack
             .push(U256::from(current_call_frame.sub_return_data.len()))?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // RETURNDATACOPY operation
     pub fn op_returndatacopy(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         let dest_offset = current_call_frame.stack.pop()?;
         let returndata_offset: usize = current_call_frame
             .stack
@@ -392,7 +392,7 @@ impl VM {
         )?;
 
         if size == 0 && returndata_offset == 0 {
-            return Ok(OpcodeSuccess::Continue);
+            return Ok(OpcodeResult::Continue);
         }
 
         let sub_return_data_len = current_call_frame.sub_return_data.len();
@@ -422,14 +422,14 @@ impl VM {
 
         memory::try_store_data(&mut current_call_frame.memory, dest_offset, &data)?;
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 
     // EXTCODEHASH operation
     pub fn op_extcodehash(
         &mut self,
         current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
+    ) -> Result<OpcodeResult, VMError> {
         let address = word_to_address(current_call_frame.stack.pop()?);
 
         let (account_info, address_was_cold) = self.access_account(address);
@@ -447,13 +447,13 @@ impl VM {
             // An account is considered empty when it has no code and zero nonce and zero balance. [EIP-161]
             if account_info.is_empty() {
                 current_call_frame.stack.push(U256::zero())?;
-                return Ok(OpcodeSuccess::Continue);
+                return Ok(OpcodeResult::Continue);
             }
 
             let hash = U256::from_big_endian(keccak(account_info.bytecode).as_fixed_bytes());
             current_call_frame.stack.push(hash)?;
         }
 
-        Ok(OpcodeSuccess::Continue)
+        Ok(OpcodeResult::Continue)
     }
 }
